Add points list and points context
Points list is used for displaying a list of saved points.

Points context us used for passing in the points and the function for setting the current points.

Points list context is used for passing in the selected points, a function to set the selected points, and a function to delete saved points.